backend/requirements.txt
fastapi[all]==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
alembic==1.13.1
asyncpg==0.29.0  # For async PostgreSQL support
